"Fix" PyPy build by using meson 0.55 and a hacky patch.
This is the last known version of meson to not automatically flag an
error when `python.dependency()` returns not found for PyPy. By
additionally patching out explicit checks for whether python was
found, the build can then go on to succeed anyway because PyPy was, in
actual effect, found. For the bug report on why `python.depencency()`
fails for PyPy in the first place, see
https://github.com/mesonbuild/meson/issues/8570